Create SafeInteriorHandle from SafeSharedX509StackHandle.
Extract the parent tracking logic to a base class to enable the generation of more SafeShared*Handle types without having to duplicate the logic.
